refactor: move all props builder to external file
