--- src/components/react_canvas/area_geometries.test.tsx
@@ -1,98 +1,0 @@
-import { buildAreaLineProps, buildAreaPointProps, buildAreaProps } from './area_geometries';
-
-describe('[canvas] Area Geometries', () => {
-  test('can build area point props', () => {
-    const props = buildAreaPointProps({
-      areaIndex: 1,
-      pointIndex: 2,
-      x: 10,
-      y: 20,
-      radius: 30,
-      strokeWidth: 2,
-      color: 'red',
-      opacity: 0.5,
-    });
-    expect(props).toEqual({
-      key: 'area-point-1-2',
-      x: 10,
-      y: 20,
-      radius: 30,
-      strokeWidth: 2,
-      strokeEnabled: true,
-      stroke: 'red',
-      fill: 'white',
-      opacity: 0.5,
-      strokeHitEnabled: false,
-      perfectDrawEnabled: false,
-      listening: false,
-    });
-
-    const propsNoStroke = buildAreaPointProps({
-      areaIndex: 1,
-      pointIndex: 2,
-      x: 10,
-      y: 20,
-      radius: 30,
-      strokeWidth: 0,
-      color: 'red',
-      opacity: 0.5,
-    });
-    expect(propsNoStroke).toEqual({
-      key: 'area-point-1-2',
-      x: 10,
-      y: 20,
-      radius: 30,
-      strokeWidth: 0,
-      strokeEnabled: false,
-      stroke: 'red',
-      fill: 'white',
-      opacity: 0.5,
-      strokeHitEnabled: false,
-      perfectDrawEnabled: false,
-      listening: false,
-    });
-  });
-  test('can build area path props', () => {
-    const props = buildAreaProps({
-      index: 1,
-      areaPath: 'M0,0L10,10Z',
-      color: 'red',
-      opacity: 0.5,
-    });
-    expect(props).toEqual({
-      key: 'area-1',
-      data: 'M0,0L10,10Z',
-      fill: 'red',
-      lineCap: 'round',
-      lineJoin: 'round',
-      opacity: 0.5,
-      strokeHitEnabled: false,
-      perfectDrawEnabled: false,
-      listening: false,
-    });
-  });
-  test('can build area line path props', () => {
-    const props = buildAreaLineProps({
-      index: 1,
-      linePath: 'M0,0L10,10Z',
-      color: 'red',
-      strokeWidth: 1,
-      geometryStyle: {
-        opacity: 0.5,
-      },
-    });
-    expect(props).toEqual({
-      key: `area-line-1`,
-      data: 'M0,0L10,10Z',
-      stroke: 'red',
-      strokeWidth: 1,
-      lineCap: 'round',
-      lineJoin: 'round',
-      opacity: 0.5,
-      strokeHitEnabled: false,
-      perfectDrawEnabled: false,
-      listening: false,
-    });
-    expect(props.fill).toBeFalsy();
-  });
-});
